+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 This file will document the most important changes for each released version, starting from version 0.4.0
 
 ## [UNRELEASED]
+
+### Features
+- Add CLI command to hexdump the data contents of an input file and exit without showing the GUI
+- Add CLI command to binary-dump the data contents of an input file and exit without showing the GUI